+++ application/modules/common/controllers/UserDefaultController.php
@@ -743,6 +743,7 @@ class UserDefaultController extends Zend_Controller_Action
                     $photoFileName = $subform->PHOTO->getFileName();
 
                     try {
+                        $user->savePhotoWithUuid($photoFileName, $user->getUuid());
                         $user->savePhoto($photoFileName);
                         Episciences_Auth::incrementPhotoVersion();
                     } catch (Exception $e) {
@@ -1439,11 +1440,15 @@ class UserDefaultController extends Zend_Controller_Action
 
         $res = false;
         if ($this->getRequest()->isXmlHttpRequest() && isset($params['uid'])) {
-            if (Episciences_Auth::getUid() == $params['uid'] || Episciences_Auth::isSecretary()) {
-                $user = new Ccsd_User_Models_User(['uid' => $params['uid']]);
-                $user->deletePhoto();
+            $uid = (int)$params['uid'];
+            if (Episciences_Auth::getUid() === $uid || Episciences_Auth::isSecretary()) {
+                $user = new Ccsd_User_Models_User(['uid' => $uid]);
+                $user->setUuid(Episciences_Auth::getUuid());
+                $user->deletePhoto($user->getUuid());
                 Episciences_Auth::incrementPhotoVersion();
-                if (Episciences_Auth::getUid() == $params['uid']) {
+                $user->deletePhoto($user->getUid());
+                Episciences_Auth::incrementPhotoVersion();
+                if (Episciences_Auth::getUid() === $uid) {
                     $res = '1';
                 } else {
                     $res = '2';
@@ -1461,7 +1466,23 @@ class UserDefaultController extends Zend_Controller_Action
     {
         $this->_helper->layout()->disableLayout();
         $this->_helper->viewRenderer->setNoRender(true);
-        $uid = $this->getParam('uid', 0);
+
+
+        $cleanedUuid = '';
+        $user = null;
+        $uuid = null;
+
+        if ($this->hasParam('uuid')){
+            $photoPathNameFromUuid = true;
+            $uuid = (string)$this->getParam('uuid');
+            $uid = Episciences_UserManager::getCorrespondingUidFromUuid($uuid);
+            $cleanedUuid = Episciences_Tools::getCleanedUuid($uuid);
+
+        } else {
+            $photoPathNameFromUuid = false;
+            $uid = $this->getParam('uid', 0);
+        }
+
         $size = $this->getParam('size', Ccsd_User_Models_User::IMG_NAME_NORMAL);
 
         $photoPathName = false;
@@ -1487,27 +1508,32 @@ class UserDefaultController extends Zend_Controller_Action
 
 
         // photo of a specific user
-        if ($uid != 0) {
+        if ($uid !== 0) {
             $user = new Ccsd_User_Models_User(['uid' => $uid]);
-            $photoPathName = $user->getPhotoPathName($size);
+
+            if($cleanedUuid !== ''){
+                $user->setUuid($uuid);
+            }
+            $photoPathName = $user->getPhotoPathName($photoPathNameFromUuid ? $cleanedUuid  : $user->getUid(), $size);
         } else {
             // nobody or logged user
             $uid = Episciences_Auth::getUid();
-            if ($uid != 0) {
+            if ($uid !== 0) {
                 $user = new Ccsd_User_Models_User(['uid' => $uid]);
-                $photoPathName = $user->getPhotoPathName($size);
+                $user->setUuid(Episciences_Auth::getUuid());
+                $photoPathName = $user->getPhotoPathName($photoPathNameFromUuid ? $cleanedUuid : $user->getUid(), $size);
             }
         }
 
         if (!$photoPathName) {
-            if ($size === Ccsd_User_Models_User::IMG_NAME_INITIALS) {
+            if ($size === Ccsd_User_Models_User::IMG_NAME_INITIALS && $user) {
 
-                $userPhotoPath = $user->getPhotoPath();
+                $userPhotoPath = $user->getPhotoPath($photoPathNameFromUuid ? $cleanedUuid : $user->getUid());
 
                 if (!is_dir($userPhotoPath) && !mkdir($userPhotoPath, 0777, true) && !is_dir($userPhotoPath)) {
                     trigger_error(sprintf('Directory "%s" was not created', $userPhotoPath), E_USER_WARNING);
                 }
-                $photoPathName = $userPhotoPath . '/' . Ccsd_User_Models_User::IMG_PREFIX_INITIALS . $user->getUid() . '.svg';
+                $photoPathName = $userPhotoPath . '/' . Ccsd_User_Models_User::IMG_PREFIX_INITIALS . $photoPathNameFromUuid ? $cleanedUuid: $user->getUid() . '.svg';
                 $data = Episciences_View_Helper_GetAvatar::asSvg($screenName);
                 file_put_contents($photoPathName, $data);